feat: Build Complete

+++ application/DotNetMeshClient/NHS.Mesh.Client/Configuration/MeshConnectConfiguration.cs
@@ -54,6 +54,6 @@ namespace NHS.MESH.Client.Configuration
         /// <summary>Gets or sets a value indicating whether proxy use default credentials.</summary>
         public bool ProxyUseDefaultCredentials { get; set; }
         /// <summary>Gets the chunk size in bytes for sending chunked messages 19Mb limit outside of HSCN 100Mb limit within</summary>
-        public long ChunkSize {get;set;}
+        public int ChunkSize {get;set;}
   }
 }
