+++ application/DotNetMeshClient/NHS.Mesh.Client/Contracts/Configurations/IMeshConnectConfiguration.cs
@@ -52,6 +52,6 @@ namespace NHS.MESH.Client.Contracts.Configurations
         /// <summary>Gets or sets a value indicating whether proxy use default credentials.</summary>
         bool ProxyUseDefaultCredentials { get; set; }
         /// <summary>Gets the chunk size in bytes for sending chunked messages 19Mb limit outside of HSCN 100Mb limit within</summary>
-        long ChunkSize { get; set; }
+        int ChunkSize { get; set; }
     }
 }
